feat: Extend TimeInterval class to support addition and subtraction of integers
The code changes in this commit extend the TimeInterval class to support the addition and subtraction of integers. Adding an integer to a TimeInterval object means adding seconds, while subtracting an integer means removing seconds. This enhancement improves the functionality of the class and allows for more flexible time interval calculations.